pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "tetra-mix", email = "ryouma1128@outlook.jp" }
]
dependencies = [
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
    "numpy>=2.2.4",
    "websockets>=15.0.1",
    "sounddevice>=0.5.1",
    "wave>=0.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/python"]
